remove modelos de Reserva, Servico, TipoServico e Veiculo; atualiza ServicosServices e UsersServices
